Added the ability to have packages inside a module map, and made the steam page a package.
