+++ src/gui/windows/steam/__init__.py
@@ -1,0 +1,1 @@
+from .steam import SteamWindow
