+++ src/utils/module_map.py
@@ -1,10 +1,13 @@
+from os import listdir
 from os.path import dirname, basename, isfile, splitext
 from glob import glob
 from inspect import isclass
 
+
 class ModuleMeta(type):
     def __getattr__(cls, item):
         return getattr(cls.map, item)
+
 
 class ModuleMap(metaclass=ModuleMeta):
     map: dict = {}
@@ -26,7 +29,7 @@ class ModuleMap(metaclass=ModuleMeta):
 
     @classmethod
     def _import_modules(cls, file, package):
-        modules = glob(dirname(file) + "/*.py")
+        modules = listdir(dirname(file))
         return [(splitext(basename(module))[0],
                  __import__('.'.join([package, splitext(basename(module))[0]]), fromlist='*'))
-                for module in modules if isfile(module) and not basename(module).startswith('__')]
+                for module in modules if not basename(module).startswith('__')]
